shop_app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.24.0
pandas>=1.3.0
google-auth>=2.3.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
openpyxl>=3.0.7
pytest>=6.2.5
python-dotenv>=0.19.0
pydantic>=1.8.2
httpx>=0.18.2
pytest-asyncio>=0.15.1